adding exec application command
