+++ src/presentation/legallead.installer.tests/ZipProgressTests.cs
@@ -1,10 +1,5 @@
 ﻿using Bogus;
 using legallead.installer.Classes;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace legallead.installer.tests
 {
