+++ src/presentation/legallead.installer/Models/DirectoryInfoModel.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace legallead.installer.Models
+﻿namespace legallead.installer.Models
 {
     public class DirectoryInfoModel
     {
